backend/requirements.txt
Flask>=3.0
flask-cors>=4.0
Flask-SQLAlchemy>=3.1
PyMySQL>=1.1
python-dotenv>=1.0
Flask-Caching>=2.0.0
flask-compress>=1.14